rustc_session::config::cfg
Type Alias Cfg

pub type Cfg = FxIndexSet<(Symbol, Option<Symbol>)>;

The parsed --cfg options that define the compilation environment of the crate, used to drive conditional compilation.

An FxIndexSet is used to ensure deterministic ordering of error messages relating to --cfg.

Aliased Type

struct Cfg { /* private fields */ }

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 56 bytes