rustc_session::config

Function rustc_optgroups

pub fn rustc_optgroups() -> Vec<RustcOptGroup>

Returns all rustc command line options, including metadata for each option, such as whether the option is stable.

Option style guidelines

<param>: Indicates a required parameter

[param]: Indicates an optional parameter

|: Indicates a mutually exclusive option

*: a list element with description