Function rustc_session::config::rustc_short_optgroups

pub fn rustc_short_optgroups() -> Vec<RustcOptGroup>

Returns the “short” subset of the rustc command line options, including metadata for each option, such as whether the option is part of the stable long-term interface for rustc.