rustc_session::config

Module native_libs

Parser for the -l command-line option, which links the generated crate to a native library.

(There is also a similar but separate syntax for #[link] attributes, which have their own parser in rustc_metadata.)

Structs

NativeLibParts 🔒

ParseNativeLibCx 🔒

Functions

parse_and_apply_modifier 🔒

Parses one of the comma-separated modifiers (prefixed by + or -), and modifies native_lib appropriately.

parse_native_lib 🔒

Parses the value of a single -l option.

parse_native_libs 🔒

Parses all -l options.

split_native_lib_value 🔒

Splits a string of the form [KIND[:MODIFIERS]=]NAME[:NEW_NAME] into those individual parts. This cannot fail, but the resulting strings require further validation.