Struct rustc_session::config::OutputTypes

pub struct OutputTypes(BTreeMap<OutputType, Option<PathBuf>>);

Use tree-based collections to cheaply get a deterministic Hash implementation. Do not switch BTreeMap out for an unsorted container type! That would break dependency tracking for command-line arguments. Also only hash keys, since tracking should only depend on the output types, not the paths they’re written to.

Tuple Fields
0: BTreeMap<OutputType, Option<PathBuf>>
Implementations

impl OutputTypes

pub fn new(entries: &[(OutputType, Option<PathBuf>)]) -> OutputTypes

pub fn get(&self, key: &OutputType) -> Option<&Option<PathBuf>>

pub fn contains_key(&self, key: &OutputType) -> bool

pub fn keys(&self) -> BTreeMapKeysIter<'_, OutputType, Option<PathBuf>>

pub fn values(&self) -> BTreeMapValuesIter<'_, OutputType, Option<PathBuf>>

pub fn len(&self) -> usize

pub fn should_codegen(&self) -> bool

Returns true if any of the output types require codegen or linking.

pub fn should_link(&self) -> bool

Returns true if any of the output types require linking.