Struct rustc_session::config::UnstableOptions

pub struct UnstableOptions {

    pub allow_features: Option<Vec<String>>,
    pub always_encode_mir: bool,
    pub assert_incr_state: Option<String>,
    pub assume_incomplete_release: bool,
    pub binary_dep_depinfo: bool,
    pub box_noalias: bool,
    pub branch_protection: Option<BranchProtection>,
    pub cf_protection: CFProtection,
    pub check_cfg_all_expected: bool,
    pub codegen_backend: Option<String>,
    pub combine_cgu: bool,
    pub coverage_options: CoverageOptions,
    pub crate_attr: Vec<String>,
    pub cross_crate_inline_threshold: InliningThreshold,
    pub debug_info_for_profiling: bool,
    pub debuginfo_compression: DebugInfoCompression,
    pub deduplicate_diagnostics: bool,
    pub default_hidden_visibility: Option<bool>,
    pub dep_info_omit_d_target: bool,
    pub direct_access_external_data: Option<bool>,
    pub dual_proc_macros: bool,
    pub dump_dep_graph: bool,
    pub dump_mir: Option<String>,
    pub dump_mir_dataflow: bool,
    pub dump_mir_dir: String,
    pub dump_mir_exclude_alloc_bytes: bool,
    pub dump_mir_exclude_pass_number: bool,
    pub dump_mir_graphviz: bool,
    pub dump_mono_stats: SwitchWithOptPath,
    pub dump_mono_stats_format: DumpMonoStatsFormat,
    pub dwarf_version: Option<u32>,
    pub dylib_lto: bool,
    pub eagerly_emit_delayed_bugs: bool,
    pub ehcont_guard: bool,
    pub emit_stack_sizes: bool,
    pub emit_thin_lto: bool,
    pub enforce_type_length_limit: bool,
    pub export_executable_symbols: bool,
    pub external_clangrt: bool,
    pub extra_const_ub_checks: bool,
    pub fewer_names: Option<bool>,
    pub fixed_x18: bool,
    pub flatten_format_args: bool,
    pub force_unstable_if_unmarked: bool,
    pub fuel: Option<(String, u64)>,
    pub function_return: FunctionReturn,
    pub function_sections: Option<bool>,
    pub future_incompat_test: bool,
    pub graphviz_dark_mode: bool,
    pub graphviz_font: String,
    pub has_thread_local: Option<bool>,
    pub hir_stats: bool,
    pub human_readable_cgu_names: bool,
    pub identify_regions: bool,
    pub ignore_directory_in_diagnostics_source_blocks: Vec<String>,
    pub incremental_ignore_spans: bool,
    pub incremental_info: bool,
    pub incremental_verify_ich: bool,
    pub inline_in_all_cgus: Option<bool>,
    pub inline_llvm: bool,
    pub inline_mir: Option<bool>,
    pub inline_mir_forwarder_threshold: Option<usize>,
    pub inline_mir_hint_threshold: Option<usize>,
    pub inline_mir_preserve_debug: Option<bool>,
    pub inline_mir_threshold: Option<usize>,
    pub input_stats: bool,
    pub instrument_mcount: bool,
    pub instrument_xray: Option<InstrumentXRay>,
    pub layout_seed: Option<u64>,
    pub link_directives: bool,
    pub link_native_libraries: bool,
    pub link_only: bool,
    pub linker_features: LinkerFeaturesCli,
    pub lint_mir: bool,
    pub llvm_module_flag: Vec<(String, u32, String)>,
    pub llvm_plugins: Vec<String>,
    pub llvm_time_trace: bool,
    pub location_detail: LocationDetail,
    pub ls: Vec<String>,
    pub macro_backtrace: bool,
    pub maximal_hir_to_mir_coverage: bool,
    pub merge_functions: Option<MergeFunctions>,
    pub meta_stats: bool,
    pub mir_emit_retag: bool,
    pub mir_enable_passes: Vec<(String, bool)>,
    pub mir_include_spans: bool,
    pub mir_keep_place_mention: bool,
    pub mir_opt_level: Option<usize>,
    pub move_size_limit: Option<usize>,
    pub mutable_noalias: bool,
    pub next_solver: Option<NextSolverConfig>,
    pub nll_facts: bool,
    pub nll_facts_dir: String,
    pub no_analysis: bool,
    pub no_codegen: bool,
    pub no_generate_arange_section: bool,
    pub no_implied_bounds_compat: bool,
    pub no_jump_tables: bool,
    pub no_leak_check: bool,
    pub no_link: bool,
    pub no_parallel_backend: bool,
    pub no_profiler_runtime: bool,
    pub no_trait_vptr: bool,
    pub no_unique_section_names: bool,
    pub normalize_docs: bool,
    pub on_broken_pipe: OnBrokenPipe,
    pub oom: OomStrategy,
    pub osx_rpath_install_name: bool,
    pub packed_bundled_libs: bool,
    pub panic_abort_tests: bool,
    pub panic_in_drop: PanicStrategy,
    pub parse_only: bool,
    pub patchable_function_entry: PatchableFunctionEntry,
    pub plt: Option<bool>,
    pub polonius: Polonius,
    pub polymorphize: bool,
    pub pre_link_arg: (),
    pub pre_link_args: Vec<String>,
    pub precise_enum_drop_elaboration: bool,
    pub print_codegen_stats: bool,
    pub print_fuel: Option<String>,
    pub print_llvm_passes: bool,
    pub print_mono_items: Option<String>,
    pub print_type_sizes: bool,
    pub print_vtable_sizes: bool,
    pub proc_macro_backtrace: bool,
    pub proc_macro_execution_strategy: ProcMacroExecutionStrategy,
    pub profile: bool,
    pub profile_closures: bool,
    pub profile_emit: Option<PathBuf>,
    pub profile_sample_use: Option<PathBuf>,
    pub profiler_runtime: String,
    pub query_dep_graph: bool,
    pub randomize_layout: bool,
    pub relax_elf_relocations: Option<bool>,
    pub remap_cwd_prefix: Option<PathBuf>,
    pub remap_path_scope: RemapPathScopeComponents,
    pub remark_dir: Option<PathBuf>,
    pub sanitizer: SanitizerSet,
    pub sanitizer_cfi_canonical_jump_tables: Option<bool>,
    pub sanitizer_cfi_generalize_pointers: Option<bool>,
    pub sanitizer_cfi_normalize_integers: Option<bool>,
    pub sanitizer_dataflow_abilist: Vec<String>,
    pub sanitizer_memory_track_origins: usize,
    pub sanitizer_recover: SanitizerSet,
    pub saturating_float_casts: Option<bool>,
    pub self_profile: SwitchWithOptPath,
    pub self_profile_counter: String,
    pub self_profile_events: Option<Vec<String>>,
    pub share_generics: Option<bool>,
    pub shell_argfiles: bool,
    pub show_span: Option<String>,
    pub simulate_remapped_rust_src_base: Option<PathBuf>,
    pub span_debug: bool,
    pub span_free_formats: bool,
    pub split_dwarf_inlining: bool,
    pub split_dwarf_kind: SplitDwarfKind,
    pub split_lto_unit: Option<bool>,
    pub src_hash_algorithm: Option<SourceFileHashAlgorithm>,
    pub stack_protector: StackProtector,
    pub staticlib_allow_rdylib_deps: bool,
    pub staticlib_prefer_dynamic: bool,
    pub strict_init_checks: bool,
    pub teach: bool,
    pub temps_dir: Option<String>,
    pub terminal_urls: TerminalUrl,
    pub thinlto: Option<bool>,
    pub threads: usize,
    pub time_llvm_passes: bool,
    pub time_passes: bool,
    pub time_passes_format: TimePassesFormat,
    pub tiny_const_eval_limit: bool,
    pub tls_model: Option<TlsModel>,
    pub trace_macros: bool,
    pub track_diagnostics: bool,
    pub translate_additional_ftl: Option<PathBuf>,
    pub translate_directionality_markers: bool,
    pub translate_lang: Option<LanguageIdentifier>,
    pub translate_remapped_path_to_local_path: bool,
    pub trap_unreachable: Option<bool>,
    pub treat_err_as_bug: Option<NonZero<usize>>,
    pub trim_diagnostic_paths: bool,
    pub tune_cpu: Option<String>,
    pub ub_checks: Option<bool>,
    pub ui_testing: bool,
    pub uninit_const_chunk_threshold: usize,
    pub unleash_the_miri_inside_of_you: bool,
    pub unpretty: Option<String>,
    pub unsound_mir_opts: bool,
    pub unstable_options: bool,
    pub use_ctors_section: Option<bool>,
    pub use_sync_unwind: Option<bool>,
    pub validate_mir: bool,
    pub verbose_asm: bool,
    pub verbose_internals: bool,
    pub verify_llvm_ir: bool,
    pub virtual_function_elimination: bool,
    pub wasi_exec_model: Option<WasiExecModel>,
    pub wasm_c_abi: WasmCAbi,
    pub write_long_types_to_disk: bool,

}

Fields
allow_features: Option<Vec<String>>always_encode_mir: boolassert_incr_state: Option<String>assume_incomplete_release: boolbinary_dep_depinfo: boolbox_noalias: boolbranch_protection: Option<BranchProtection>cf_protection: CFProtectioncheck_cfg_all_expected: boolcodegen_backend: Option<String>combine_cgu: boolcoverage_options: CoverageOptionscrate_attr: Vec<String>cross_crate_inline_threshold: InliningThresholddebug_info_for_profiling: booldebuginfo_compression: DebugInfoCompressiondeduplicate_diagnostics: booldefault_hidden_visibility: Option<bool>dep_info_omit_d_target: booldirect_access_external_data: Option<bool>dual_proc_macros: booldump_dep_graph: booldump_mir: Option<String>dump_mir_dataflow: booldump_mir_dir: Stringdump_mir_exclude_alloc_bytes: booldump_mir_exclude_pass_number: booldump_mir_graphviz: booldump_mono_stats: SwitchWithOptPathdump_mono_stats_format: DumpMonoStatsFormatdwarf_version: Option<u32>dylib_lto: booleagerly_emit_delayed_bugs: boolehcont_guard: boolemit_stack_sizes: boolemit_thin_lto: boolenforce_type_length_limit: boolexport_executable_symbols: boolexternal_clangrt: boolextra_const_ub_checks: boolfewer_names: Option<bool>fixed_x18: boolflatten_format_args: boolforce_unstable_if_unmarked: boolfuel: Option<(String, u64)>function_return: FunctionReturnfunction_sections: Option<bool>future_incompat_test: boolgraphviz_dark_mode: boolgraphviz_font: Stringhas_thread_local: Option<bool>hir_stats: boolhuman_readable_cgu_names: boolidentify_regions: boolignore_directory_in_diagnostics_source_blocks: Vec<String>incremental_ignore_spans: boolincremental_info: boolincremental_verify_ich: boolinline_in_all_cgus: Option<bool>inline_llvm: boolinline_mir: Option<bool>inline_mir_forwarder_threshold: Option<usize>inline_mir_hint_threshold: Option<usize>inline_mir_preserve_debug: Option<bool>inline_mir_threshold: Option<usize>input_stats: boolinstrument_mcount: boolinstrument_xray: Option<InstrumentXRay>layout_seed: Option<u64>link_directives: boollink_native_libraries: boollink_only: boollinker_features: LinkerFeaturesClilint_mir: boolllvm_module_flag: Vec<(String, u32, String)>llvm_plugins: Vec<String>llvm_time_trace: boollocation_detail: LocationDetaills: Vec<String>macro_backtrace: boolmaximal_hir_to_mir_coverage: boolmerge_functions: Option<MergeFunctions>meta_stats: boolmir_emit_retag: boolmir_enable_passes: Vec<(String, bool)>mir_include_spans: boolmir_keep_place_mention: boolmir_opt_level: Option<usize>move_size_limit: Option<usize>mutable_noalias: boolnext_solver: Option<NextSolverConfig>nll_facts: boolnll_facts_dir: Stringno_analysis: boolno_codegen: boolno_generate_arange_section: boolno_implied_bounds_compat: boolno_jump_tables: boolno_leak_check: boolno_link: boolno_parallel_backend: boolno_profiler_runtime: boolno_trait_vptr: boolno_unique_section_names: boolnormalize_docs: boolon_broken_pipe: OnBrokenPipeoom: OomStrategyosx_rpath_install_name: boolpacked_bundled_libs: boolpanic_abort_tests: boolpanic_in_drop: PanicStrategyparse_only: boolpatchable_function_entry: PatchableFunctionEntryplt: Option<bool>polonius: Poloniuspolymorphize: boolpre_link_arg: ()pre_link_args: Vec<String>precise_enum_drop_elaboration: boolprint_codegen_stats: boolprint_fuel: Option<String>print_llvm_passes: boolprint_mono_items: Option<String>print_type_sizes: boolprint_vtable_sizes: boolproc_macro_backtrace: boolproc_macro_execution_strategy: ProcMacroExecutionStrategyprofile: boolprofile_closures: boolprofile_emit: Option<PathBuf>profile_sample_use: Option<PathBuf>profiler_runtime: Stringquery_dep_graph: boolrandomize_layout: boolrelax_elf_relocations: Option<bool>remap_cwd_prefix: Option<PathBuf>remap_path_scope: RemapPathScopeComponentsremark_dir: Option<PathBuf>sanitizer: SanitizerSetsanitizer_cfi_canonical_jump_tables: Option<bool>sanitizer_cfi_generalize_pointers: Option<bool>sanitizer_cfi_normalize_integers: Option<bool>sanitizer_dataflow_abilist: Vec<String>sanitizer_memory_track_origins: usizesanitizer_recover: SanitizerSetsaturating_float_casts: Option<bool>self_profile: SwitchWithOptPathself_profile_counter: Stringself_profile_events: Option<Vec<String>>

keep this in sync with the event filter names in librustc_data_structures/profiling.rs

share_generics: Option<bool>shell_argfiles: boolshow_span: Option<String>simulate_remapped_rust_src_base: Option<PathBuf>span_debug: boolspan_free_formats: bool

o/w tests have closure@path

split_dwarf_inlining: boolsplit_dwarf_kind: SplitDwarfKindsplit_lto_unit: Option<bool>src_hash_algorithm: Option<SourceFileHashAlgorithm>stack_protector: StackProtectorstaticlib_allow_rdylib_deps: boolstaticlib_prefer_dynamic: boolstrict_init_checks: boolteach: booltemps_dir: Option<String>terminal_urls: TerminalUrlthinlto: Option<bool>threads: usize

We default to 1 here since we want to behave like a sequential compiler for now. This’ll likely be adjusted in the future. Note that -Zthreads=0 is the way to get the num_cpus behavior.

time_llvm_passes: booltime_passes: booltime_passes_format: TimePassesFormattiny_const_eval_limit: booltls_model: Option<TlsModel>trace_macros: booltrack_diagnostics: booltranslate_additional_ftl: Option<PathBuf>translate_directionality_markers: booltranslate_lang: Option<LanguageIdentifier>translate_remapped_path_to_local_path: booltrap_unreachable: Option<bool>treat_err_as_bug: Option<NonZero<usize>>trim_diagnostic_paths: booltune_cpu: Option<String>ub_checks: Option<bool>ui_testing: booluninit_const_chunk_threshold: usizeunleash_the_miri_inside_of_you: boolunpretty: Option<String>unsound_mir_opts: boolunstable_options: bool

This name is kind of confusing: Most unstable options enable something themselves, while this just allows “normal” options to be feature-gated.

use_ctors_section: Option<bool>use_sync_unwind: Option<bool>validate_mir: boolverbose_asm: boolverbose_internals: boolverify_llvm_ir: boolvirtual_function_elimination: boolwasi_exec_model: Option<WasiExecModel>wasm_c_abi: WasmCAbiwrite_long_types_to_disk: bool
Implementations

impl UnstableOptions

pub fn dcx_flags(&self, can_emit_warnings: bool) -> DiagCtxtFlags

pub fn src_hash_algorithm(&self, target: &Target) -> SourceFileHashAlgorithm

impl UnstableOptions

pub fn build(early_dcx: &EarlyDiagCtxt, matches: &Matches) -> UnstableOptions

fn dep_tracking_hash( &self, for_crate_hash: bool, error_format: ErrorOutputType, ) -> u64