Trait rustc_session::cstore::MetadataLoader

pub trait MetadataLoader {
    fn get_rlib_metadata(
        &self,
        target: &Target,
        filename: &Path
    ) -> Result<MetadataRef, String>;
    fn get_dylib_metadata(
        &self,
        target: &Target,
        filename: &Path
    ) -> Result<MetadataRef, String>;
}

The backend’s way to give the crate store access to the metadata in a library. Note that it returns the raw metadata bytes stored in the library file, whether it is compressed, uncompressed, some weird mix, etc. rmeta files are backend independent and not handled here.

At the time of this writing, there is only one backend and one way to store metadata in library – this trait just serves to decouple rustc_metadata from the archive reader, which depends on LLVM.

Required Methods

fn get_rlib_metadata(
&self,
target: &Target,
filename: &Path
) -> Result<MetadataRef, String>

fn get_dylib_metadata(
&self,
target: &Target,
filename: &Path
) -> Result<MetadataRef, String>