rustc_session
Module output

Related to out filenames of compilation (e.g. binaries).

Constants

CRATE_TYPES

Functions

categorize_crate_type

check_file_is_writeable

Make sure files are writeable. Mac, FreeBSD, and Windows system linkers check this already – however, the Linux linker will happily overwrite a read-only file. We should be consistent.

collect_crate_types

default_output_for_target

Returns default crate type for target

filename_for_input

filename_for_metadata

find_crate_name

invalid_output_for_target

Checks if target supports crate_type as output

is_writeable 🔒

out_filename

validate_crate_name