rustc_session::parse

Function feature_err

pub fn feature_err(
    sess: &Session,
    feature: Symbol,
    span: impl Into<MultiSpan>,
    explain: impl Into<DiagMessage>,
) -> Diag<'_>

Construct a diagnostic for a language feature error due to the given span. The feature’s Symbol is the one you used in unstable.rs and rustc_span::symbol.