Function rustc_session::parse::feature_warn_issue

pub fn feature_warn_issue(
    sess: &ParseSess,
    feature: Symbol,
    span: Span,
    issue: GateIssue,
    explain: &'static str
)

Construct a future incompatibility diagnostic for a feature gate.

This diagnostic is only a warning and does not cause compilation to fail.

This variant allows you to control whether it is a library or language feature. Almost always, you want to use this for a language feature. If so, prefer feature_warn.