Trait rustc_session::RemapFileNameExt

pub trait RemapFileNameExt {
    type Output<'a>
       where Self: 'a;

    // Required methods
    fn for_scope(
        &self,
        sess: &Session,
        scope: RemapPathScopeComponents
    ) -> Self::Output<'_>;
    fn for_codegen(&self, sess: &Session) -> Self::Output<'_>;
}

Required Associated Types

type Output<'a> where Self: 'a

Required Methods

fn for_scope( &self, sess: &Session, scope: RemapPathScopeComponents ) -> Self::Output<'_>

Returns a possibly remapped filename based on the passed scope and remap cli options.

One and only one scope should be passed to this method. For anything related to “codegen” see the RemapFileNameExt::for_codegen method.

fn for_codegen(&self, sess: &Session) -> Self::Output<'_>

Return a possibly remapped filename, to be used in “codegen” related parts.

Object Safety

This trait is not object safe.