rustc_session
Module cstore

the rustc crate store interface. This also includes types that are mostly used as a part of that interface, but these should probably get a better home if someone can find one.

Structs

CrateSource

Where a crate came from on the local filesystem. One of these three options must be non-None.

DllImport

ExternCrate

ForeignModule

NativeLib

Untracked

Enums

CrateDepKind

DllCallingConvention

Calling convention for a function defined in an external library.

ExternCrateSource

LinkagePreference

PeImportNameType

Different ways that the PE Format can decorate a symbol name. From https://docs.microsoft.com/en-us/windows/win32/debug/pe-format#import-name-type

Traits

CrateStore

A store of Rust crates, through which their metadata can be accessed.

Type Aliases

CrateStoreDyn