Module rustc_session::filesearch

A module for searching for libraries

Structs

FileSearch

Functions

current_dll_path 🔒

get_or_default_sysroot

This function checks if sysroot is found using env::args().next(), and if it is not found, finds sysroot from current rustc_driver dll.

make_target_lib_path

materialize_sysroot

Returns the provided sysroot or calls get_or_default_sysroot if it’s none. Panics if get_or_default_sysroot returns an error.

sysroot_candidates