rustc_session
Module options

Modules

cgopts 🔒

dbopts 🔒

desc 🔒

parse

Macros

hash_opt 🔒

hash_substruct 🔒

insert 🔒

options 🔒

Defines all CodegenOptions/DebuggingOptions fields and parsers all at once. The goal of this macro is to define an interface that can be programmatically used by the option parser to initialize the struct without hardcoding field names all over the place.

redirect_field 🔒

top_level_options 🔒

Structs

CodegenOptions

Options

The top-level command-line options struct.

UnstableOptions

Constants

CG_OPTIONS

Z_OPTIONS

Functions

build_options 🔒

Type Aliases

OptionDescrs 🔒

OptionSetter 🔒