rustc_session

Module options

Modules

cgopts 🔒

dbopts 🔒

desc 🔒

parse

Macros

gather_tmods 🔒

gather_tmods_top_level 🔒

hash_opt 🔒

hash_substruct 🔒

insert 🔒

options 🔒

Defines all CodegenOptions/DebuggingOptions fields and parsers all at once. The goal of this macro is to define an interface that can be programmatically used by the option parser to initialize the struct without hardcoding field names all over the place.

redirect_field 🔒

tmod_enum 🔒

tmod_enum_opt 🔒

tmod_push 🔒

top_level_options 🔒

top_level_tmod_enum 🔒

Macro for generating OptionsTargetsModifiers top-level enum with impl. Will generate something like:

Structs

CodegenOptions

ExtendedTargetModifierInfo

Extended target modifier info. For example, when external target modifier is ‘-Zregparm=2’: Target modifier enum value + user value (‘2’) from external crate is converted into description: prefix (‘Z’), name (‘regparm’), tech value (‘Some(2)’).

OptionDesc

Options

The top-level command-line options struct.

TargetModifier

A recorded -Zopt_name=opt_value (or -Copt_name=opt_value) which alter the ABI or effectiveness of exploit mitigations.

UnstableOptions

Enums

CodegenOptionsTargetModifiers

OptionsTargetModifiers

UnstableOptionsTargetModifiers

Constants

CG_OPTIONS

Z_OPTIONS

Functions

build_options 🔒

tmod_push_impl 🔒

Type Aliases

OptionDescrs 🔒

OptionSetter 🔒