Module rustc_session::parse

Contains ParseSess which holds state living beyond what one Parser might. It also serves as an input to the parser itself.

Structs

GatedSpans

Collected spans during parsing for places where a certain feature was used and should be feature gated accordingly in check_crate.

ParseSess

Info about a parsing session.

SymbolGallery

Functions

add_feature_diagnostics

Adds the diagnostics for a feature to an existing error.

add_feature_diagnostics_for_issue

Adds the diagnostics for a feature to an existing error.

feature_err

Construct a diagnostic for a language feature error due to the given span. The feature’s Symbol is the one you used in active.rs and rustc_span::symbols.

feature_err_issue

Construct a diagnostic for a feature gate error.

feature_warn

Construct a future incompatibility diagnostic for a feature gate.

feature_warn_issue

Construct a future incompatibility diagnostic for a feature gate.

Type Definitions

CrateCheckConfig

CrateConfig

The set of keys (and, optionally, values) that define the compilation environment of the crate, used to drive conditional compilation.