rustc_session

Module parse

Contains ParseSess which holds state living beyond what one Parser might. It also serves as an input to the parser itself.

Structs

GatedSpans

Collected spans during parsing for places where a certain feature was used and should be feature gated accordingly in check_crate.

ParseSess

Info about a parsing session.

SymbolGallery

Functions

add_feature_diagnostics

Adds the diagnostics for a feature to an existing error. Must be a language feature!

add_feature_diagnostics_for_issue

Adds the diagnostics for a feature to an existing error.

feature_err

Construct a diagnostic for a language feature error due to the given span. The feature’s Symbol is the one you used in unstable.rs and rustc_span::symbol.

feature_err_issue

Construct a diagnostic for a feature gate error.

feature_err_unstable_feature_bound

This is only used by unstable_feature_bound as it does not have issue number information for now. This is basically the same as feature_err_issue but without the feature issue note. If we can do a lookup for issue number from feature name, then we should directly use feature_err_issue for ambiguity error of #[unstable_feature_bound].

feature_warn

Construct a future incompatibility diagnostic for a feature gate.

feature_warn_issue

Construct a future incompatibility diagnostic for a feature gate.