rustc_session
Module session

Re-exports

pub use crate::code_stats::DataTypeKind;

pub use crate::code_stats::FieldInfo;

pub use crate::code_stats::FieldKind;

pub use crate::code_stats::SizeKind;

pub use crate::code_stats::VariantInfo;

Structs

CompilerIO

EarlyDiagCtxt

A wrapper around an DiagCtxt that is used for early error emissions.

Limit

New-type wrapper around usize for representing limits. Ensures that comparisons against limits are consistent throughout the compiler.

Limits

Session

Represents the data associated with a compilation session for a single crate.

StableCrateId

A StableCrateId is a 64-bit hash of a crate name, together with all -Cmetadata arguments, and some other data. It is to CrateNum what DefPathHash is to DefId. It is stable across compilation sessions.

Enums

CodegenUnits

CtfeBacktrace

The behavior of the CTFE engine when an error occurs with regards to backtraces.

IncrCompSession 🔒

Holds data on the current incremental compilation session, if there is one.

MetadataKind

Traits

LintStoreMarker

RemapFileNameExt

Functions

build_session

default_emitter 🔒

mk_emitter 🔒

validate_commandline_args_with_session_available 🔒

Validate command line arguments with a Session.