rustc_session::session
Trait RemapFileNameExt

pub trait RemapFileNameExt {
    type Output<'a>
       where Self: 'a;

    // Required method
    fn for_scope(
        &self,
        sess: &Session,
        scope: RemapPathScopeComponents,
    ) -> Self::Output<'_>;
}

Required Associated Types

type Output<'a> where Self: 'a

Required Methods

fn for_scope( &self, sess: &Session, scope: RemapPathScopeComponents, ) -> Self::Output<'_>

Returns a possibly remapped filename based on the passed scope and remap cli options.

One and only one scope should be passed to this method, it will panic otherwise.

Object Safety

This trait is not object safe.